debug why poly3 is not idempotent
need to just do an example by hand, unfortunately. this is the smallest polynomial, and it should really be idempotent. all one has to do is factor it as a word in the generators of choice (it doesn't matter which 2-cycle and n-cycle you pick, it appears), and then replace those generators by letters, then do the multiplication and simplify two times - once with the group elements, the other with the words in the free group, and see what possibly is going wrong.